move webwork to generate module
